refactor(mcp): separate registry from server implementations
