+++ .github/workflows/deploy-cloudflare-production.yml
@@ -287,16 +287,11 @@ jobs:
         working-directory: ./media.pollinations.ai
 
   # A Worker must exist before Gen can deploy a service binding to it. Deploy
-  # changed private MCP applications after Media (their optional upload
-  # dependency) and before Gen. The automatic Applications workflow excludes
-  # MCP Workers so it cannot race this dependency chain.
+  # changed private MCP applications before Gen. The automatic Applications
+  # workflow excludes MCP Workers so it cannot race this dependency chain.
   deploy-mcp-apps:
-    needs: [changes, deploy-media]
-    if: >-
-      always() &&
-      needs.changes.outputs.mcp_apps != '[]' &&
-      (needs.deploy-media.result == 'success' ||
-       needs.deploy-media.result == 'skipped')
+    needs: [changes]
+    if: needs.changes.outputs.mcp_apps != '[]'
     runs-on: ubuntu-latest
     strategy:
       fail-fast: false
