--- apps/ffmpeg-mcp/deploy.json
@@ -1,13 +1,0 @@
-{
-  "target": "worker",
-  "credentials": "myceli",
-  "docker": true,
-  "install": "npm ci",
-  "build": "npm test",
-  "deploy": "npm run deploy",
-  "watch": [
-    "shared/mcp-usage.ts",
-    "shared/registry/mcp.ts",
-    "shared/user-media-url.ts"
-  ]
-}
