--- apps/ffmpeg-mcp/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "pollinations-ffmpeg-mcp-worker",
-  "version": "0.1.0",
-  "private": true,
-  "type": "module",
-  "scripts": {
-    "test": "node --test worker.test.js",
-    "check": "wrangler deploy --dry-run --containers-rollout=none",
-    "deploy": "wrangler deploy",
-    "deploy:staging": "wrangler deploy --env staging"
-  },
-  "dependencies": {
-    "@cloudflare/containers": "^0.3.7",
-    "@modelcontextprotocol/server": "^2.0.0",
-    "zod": "^4.4.3"
-  },
-  "devDependencies": {
-    "@modelcontextprotocol/client": "^2.0.0",
-    "wrangler": "^4.124.0"
-  }
-}
