--- apps/ffmpeg-mcp/wrangler.jsonc
@@ -1,53 +1,0 @@
-{
-  "$schema": "./node_modules/wrangler/config-schema.json",
-  "name": "pollinations-ffmpeg-mcp",
-  "account_id": "b6ec751c0862027ba269faf7029b2501",
-  "main": "entrypoint.js",
-  "workers_dev": false,
-  "compatibility_date": "2026-08-20",
-  "compatibility_flags": ["nodejs_compat"],
-  "containers": [
-    {
-      "class_name": "FfmpegContainer",
-      "image": "./Dockerfile",
-      "max_instances": 100,
-      "instance_type": "basic"
-    }
-  ],
-  "durable_objects": {
-    "bindings": [{ "name": "FFMPEG", "class_name": "FfmpegContainer" }]
-  },
-  "migrations": [
-    { "tag": "v1", "new_sqlite_classes": ["FfmpegContainer"] }
-  ],
-  "services": [
-    { "binding": "MEDIA", "service": "pollinations-media-prod", "entrypoint": "MediaUpload" }
-  ],
-  "observability": { "enabled": true },
-  "env": {
-    "staging": {
-      "name": "pollinations-ffmpeg-mcp-staging",
-      "workers_dev": false,
-      "containers": [
-        {
-          "class_name": "FfmpegContainer",
-          "image": "./Dockerfile",
-          "max_instances": 10,
-          "instance_type": "basic"
-        }
-      ],
-      "durable_objects": {
-        "bindings": [
-          { "name": "FFMPEG", "class_name": "FfmpegContainer" }
-        ]
-      },
-      "migrations": [
-        { "tag": "v1", "new_sqlite_classes": ["FfmpegContainer"] }
-      ],
-      "services": [
-        { "binding": "MEDIA", "service": "pollinations-media-staging", "entrypoint": "MediaUpload" }
-      ],
-      "observability": { "enabled": true }
-    }
-  }
-}
