+++ gen.pollinations.ai/src/userImage.ts
@@ -2,7 +2,7 @@ import type { ImageInputErrorCode } from "@shared/error.ts";
 import { HttpError } from "@shared/http-error.ts";
 import { detectImageMimeType } from "@shared/image-mime.ts";
 import { readResponseBytes } from "@shared/response-bytes.ts";
-import { validateUserMediaUrl } from "@shared/user-media-url.ts";
+import { validateUserMediaUrl } from "./utils/user-media-url.ts";
 
 /**
  * A user-supplied image we could not use.
