+++ gen.pollinations.ai/vitest.config.ts
@@ -161,62 +161,6 @@ export default defineWorkersConfig(async ({ mode }) => {
                                     },
                                 });
                             },
-                            FFMPEG_MCP: async (request: Request) => {
-                                if (
-                                    request.headers.has("authorization") ||
-                                    request.headers.has("cookie")
-                                ) {
-                                    return new Response(
-                                        "Caller credentials reached MCP",
-                                        { status: 500 },
-                                    );
-                                }
-                                const payload = (await request.json()) as {
-                                    jsonrpc: string;
-                                    id?: string | number;
-                                    method?: string;
-                                };
-                                const headers = new Headers({
-                                    "Content-Type": "application/json",
-                                });
-                                if (payload.method === "tools/call") {
-                                    headers.set(
-                                        "x-pollinations-mcp-cost",
-                                        "0.25",
-                                    );
-                                    headers.set(
-                                        "x-pollinations-mcp-tool",
-                                        "runFfmpeg",
-                                    );
-                                    headers.set(
-                                        "x-pollinations-mcp-status",
-                                        "200",
-                                    );
-                                    headers.set(
-                                        "x-pollinations-mcp-adjustment-id",
-                                        "cloudflare.container.basic_runtime.v1",
-                                    );
-                                    headers.set(
-                                        "x-pollinations-mcp-adjustment-units",
-                                        "1",
-                                    );
-                                }
-                                return Response.json(
-                                    {
-                                        jsonrpc: payload.jsonrpc,
-                                        id: payload.id,
-                                        result: {
-                                            content: [
-                                                {
-                                                    type: "text",
-                                                    text: "proxied",
-                                                },
-                                            ],
-                                        },
-                                    },
-                                    { headers },
-                                );
-                            },
                         },
                     },
                 },
